+++ tests/test_forms.py
@@ -164,125 +164,6 @@ class FilterSetFormTests(TestCase):
         self.assertTrue(f.is_bound)
         self.assertEqual(f.data, {'title': 'Some book'})
 
-    def test_ordering(self):
-        class F(FilterSet):
-            class Meta:
-                model = User
-                fields = ['username', 'status']
-                order_by = ['status']
-
-        f = F().form
-        self.assertEqual(len(f.fields), 3)
-        self.assertIn('o', f.fields)
-        self.assertEqual(f.fields['o'].choices, [('status', 'Status')])
-
-    def test_ordering_uses_all_fields(self):
-        class F(FilterSet):
-            class Meta:
-                model = User
-                fields = ['username', 'status']
-                order_by = True
-
-        f = F().form
-        self.assertEqual(f.fields['o'].choices, [
-            ('username', 'Username'),
-            ('-username', 'Username (descending)'),
-            ('status', 'Status'),
-            ('-status', 'Status (descending)')])
-
-    def test_ordering_uses_filter_label(self):
-        class F(FilterSet):
-            username = CharFilter(label='Account')
-
-            class Meta:
-                model = User
-                fields = ['username', 'status']
-                order_by = True
-
-        f = F().form
-        self.assertEqual(f.fields['o'].choices, [
-            ('username', 'Account'),
-            ('-username', 'Account (descending)'),
-            ('status', 'Status'),
-            ('-status', 'Status (descending)')])
-
-    def test_ordering_uses_explicit_filter_name(self):
-        class F(FilterSet):
-            account = CharFilter(name='username')
-
-            class Meta:
-                model = User
-                fields = ['account', 'status']
-                order_by = True
-
-        f = F().form
-        self.assertEqual(f.fields['o'].choices, [
-            ('account', 'Account'),
-            ('-account', 'Account (descending)'),
-            ('status', 'Status'),
-            ('-status', 'Status (descending)')])
-
-    def test_ordering_with_overridden_field_name(self):
-        """
-        Set the `order_by_field` on the filterset and ensure that the
-        field name is respected.
-        """
-        class F(FilterSet):
-            class Meta:
-                model = User
-                fields = ['username', 'status']
-                order_by = ['status']
-                order_by_field = 'order'
-
-        f = F().form
-        self.assertNotIn('o', f.fields)
-        self.assertIn('order', f.fields)
-        self.assertEqual(f.fields['order'].choices, [('status', 'Status')])
-
-    def test_ordering_with_overridden_field_name_and_descending(self):
-        """
-        Set the `order_by_field` on the filterset and ensure that the
-        field name is respected.
-        """
-        class F(FilterSet):
-            class Meta:
-                model = User
-                fields = ['username', 'status']
-                order_by = ['status', '-status']
-                order_by_field = 'order'
-
-        f = F().form
-        self.assertNotIn('o', f.fields)
-        self.assertIn('order', f.fields)
-        self.assertEqual(f.fields['order'].choices, [('status', 'Status'), ('-status', 'Status (descending)')])
-
-    def test_ordering_with_overridden_field_name_and_using_all_fields(self):
-        class F(FilterSet):
-            class Meta:
-                model = User
-                fields = ['username', 'status']
-                order_by = True
-                order_by_field = 'order'
-
-        f = F().form
-        self.assertIn('order', f.fields)
-        self.assertEqual(f.fields['order'].choices, [
-            ('username', 'Username'),
-            ('-username', 'Username (descending)'),
-            ('status', 'Status'),
-            ('-status', 'Status (descending)')])
-
-    def test_ordering_with_custom_display_names(self):
-        class F(FilterSet):
-            class Meta:
-                model = User
-                fields = ['username', 'status']
-                order_by = [('status', 'Current status')]
-
-        f = F().form
-        self.assertEqual(
-            f.fields['o'].choices, [('status', 'Current status')])
-
     def test_limit_choices_to(self):
         User.objects.create(username='inactive', is_active=False, status=REGULAR)
         User.objects.create(username='active', is_active=True, status=REGULAR)
